Revert "Misc Additions"
This reverts commit 2c3b86513e7d1c80c3bbde77711789d1ff590173.

+++ Assets/Resources/Scripts/Game/Bullet.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class Bullet : MonoBehaviour {
 
 	public BulletType type;
-	public float speed;
+	public int speed;
 	public int damage;
 	public Faction faction;
 	public bool onHit;
